+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Types for gtfs-to-html files
 - Added primary key info for attributions.txt, fare_rules.txt, location_group_stops.txt, stop_areas.txt and timeframes.txt
 
+### Fixed
+- Default logging verbose = true
+
 ## [4.15.6] - 2024-11-05
 
 ### Updated